tools: Autotuner: Move part of the code to utils
